requirements.txt
amqp==5.3.1
asgiref==3.11.0
asttokens==3.0.1
billiard==4.2.3
black==25.11.0
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
decorator==5.2.1
Django==5.2.8
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-simple-history==3.10.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
executing==2.2.1
graphviz==0.21
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.2.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==9.0.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==4.6.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
xlsxwriter==3.2.9
